app/worker/requirements.txt
# Temporal SDK
temporalio

# Data processing and ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
yfinance>=0.2.0

# RAG and embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# Web scraping and HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0

# Timezone handling
pytz>=2023.3

# Database (shared with API)
psycopg2-binary>=2.9.0

# Optional: For better ML models
# xgboost>=1.7.0
# lightgbm>=4.0.0
# tensorflow>=2.13.0 